requirements.txt
fastapi
uvicorn
numpy
scikit-learn
torch
esm
numpy
pandas
fastparquet
biopython